Updating the Observer Pattern.
Instead of passing the parameters in the update function, I had updated the function signature so that simply we can get data from getter function of the observer. This will be helpful when new values will be added and we will no longer need to change the update signature everywhere.